% Generated by roxygen2 (4.0.1): do not edit by hand
\name{aggregateRanks}
\alias{aggregateRanks}
\title{Auxiliary targetPredictor functions}
\usage{
aggregateRanks(ranks, n_valid_srcs, min_src, method = "min", promote = TRUE)
}
\arguments{
\item{ranks}{data.frame with ordered scored}

\item{n_valid_srcs}{number of valid sources in the dataset}

\item{min_srcs}{minimum acceptable number fo sources}

\item{method}{'min','max', or 'geom'}

\item{promote}{add weights to improve accuracy of the method, default TRUE}
}
\value{
data.frame object with aggregate ranks
}
\description{
Auxiliary targetPredictor functions
}
\details{
This function performs aggregation phase of target prediction in getPredictedTargets()
}
\author{
Maciej Pajak \email{m.pajak@sms.ed.ac.uk}
}

