% Generated by roxygen2 (4.0.0): do not edit by hand
\name{merge_rename}
\alias{merge_rename}
\title{Auxiliary internal TargetPredictor functions}
\usage{
merge_rename(...)
}
\arguments{
\item{...}{parameters to merge}
}
\description{
This function is not meant to be called directly by the user
}
\author{
Maciej Pajak \email{m.pajak@sms.ed.ac.uk}
}

