\name{targetPredictor-package}
\alias{targetPredictor-package}
\alias{targetPredictor}
\docType{package}
\title{
targetPredictor
}
\description{
package allows aggregation of the results of multiple target prediction algorithms
}
\details{
\tabular{ll}{
Package: \tab targetPredictor\cr
Type: \tab Package\cr
Version: \tab 0.99\cr
Date: \tab 2013-10-31\cr
License: \tab GPL 2\cr
}
An overview of how to use the package, including the most important functions
}
\author{
Maintainer: Maciej Pajak <m.pajak@sms.ed.ac.uk>
}
\references{
Literature or other references for background information
}

\seealso{
some other packages?
}
\examples{
2+2
}
