% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{targetPredictor}
\alias{targetPredictor}
\alias{targetPredictor-package}
\title{Integrating miRNA target predictions from multiple sources}
\description{
Integrating miRNA target predictions from multiple sources
}
\details{
It is a package with tools to facilitate implementation of workflows requiring miRNA prediction through access to multiple prediction results (DIANA, Targetscan, PicTar and Miranda) and their aggregation.
Three aggregation methods are available: minimum, maximum and geometric mean, additional parameters provide further tuning of the results.
Predictions are available for Homo sapiens, Mus musculus and Rattus norvegicus (the last one through homology translation).
}
\examples{
#direct targets in mouse aggregated from all sources:
targets_mouse <- getPredictedTargets('let-7a',species='mmu', method='geom')
#homology-translated targets in rat aggregated from all sources
targets_rat <- getPredictedTargets('let-7a',species='mmu', method='geom')
}
\author{
Maciej Pajak \email{m.pajak@sms.ed.ac.uk}, Ian Simpson
}

